[FEAT] 두 페이지로 구현된 탭메뉴 이동시 스크롤 유지 및 플리커 현상 제거
