+++ client/src/components/NavLink.tsx
@@ -5,10 +5,12 @@ export default function NavLink({
   href,
   children,
   className,
+  scroll,
 }: {
   href: string;
   children: React.ReactNode;
   className?: string;
+  scroll?: boolean;
 }) {
   const { asPath } = useRouter();
 
@@ -23,7 +25,7 @@ export default function NavLink({
   }
 
   return (
-    <Link href={href}>
+    <Link href={href} scroll={scroll}>
       <a className={className}>{children}</a>
     </Link>
   );
